sdk/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
# This is the main SDK package name
name = "codon_sdk"
version = "0.1.0a4"
license = {text = "Apache-2.0"}
authors = [
  { name="Codon, Inc.", email="martin@codonops.ai" },
  { name="Martin Arroyo", email="martin@codonops.ai" },
  { name="Sheldon Gilbert", email="sheldon@codonops.ai"},
]
maintainers = [
  { name="Martin Arroyo", email="martin@codonops.ai" },
  { name="Carl Gordon", email="carl@codonops.ai"},
  { name="Luis Moreno", email="luis@codonops.ai"},
]
description = "The core SDK for Codon project components, including agents and data schemas."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
]
# Core dependencies needed by every part of the SDK
dependencies = [
    "opentelemetry-api",
    "opentelemetry-exporter-otlp-proto-grpc",
    "opentelemetry-sdk",
    "pydantic",
]

# This tells setuptools to find the package code inside the 'src' directory
[tool.setuptools.packages.find]
where = ["src"]

# This section defines the "extras" for optional features of the core SDK
# [project.optional-dependencies]
# Defines `pip install codon-sdk[nodespec]`
# nodespec = []

# Defines `pip install codon-sdk[agents]`
# List any dependencies needed ONLY by the agent framework here.
# agents = [
#    "dependency-for-agents==1.2.3",
# ]

[project.urls]
"Homepage" = "https://github.com/Codon-Ops/codon-sdk"
"Bug Tracker" = "https://github.com/Codon-Ops/codon-sdk/issues"
"Documentation" = "https://codon-ops.github.io/codon-sdk/"
